Fixed parallel case and ASE

+++ mala/targets/density.py
@@ -1002,7 +1002,8 @@ class Density(Target):
             barrier()
 
             # Right after setup we can delete the file.
-            os.remove(tem_input_name)
+            if get_rank() == 0:
+                os.remove(tem_input_name)
 
             printout(
                 "Total energy module: Time used by total energy initialization: {:.8f}s".format(
